backend/requirements.txt
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
distro==1.9.0
exceptiongroup==1.2.0
Flask==3.0.2
Flask-Cors==4.0.0
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
importlib_resources==6.3.2
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
openai==1.14.1
outcome==1.3.0.post0
packaging==24.0
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.8.0
PySocks==1.7.1
python-dotenv==1.0.1
python-http-client==3.3.7
pytz==2024.1
referencing==0.34.0
requests==2.31.0
rpds-py==0.18.0
selenium==4.18.1
sendgrid==6.11.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.28
starkbank-ecdsa==2.2.0
tqdm==4.66.2
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.10.0
urllib3==2.2.1
webdriver-manager==4.0.1
Werkzeug==3.0.1
wsproto==1.2.0

